+++ pinn_spm_param/postProcess/plotData.py
@@ -11,7 +11,7 @@ from keras.backend import set_floatx
 from myNN import *
 from plotsUtil_batt import *
 
-set_floatx("float64")
+set_floatx("float32")
 
 print("\n\nINFO: PLOTTING DATA OBTAINED FROM FINITE DIFFERENCE\n\n")
