+++ pinn_spm_param/postProcess/plotPINNResult_movie.py
@@ -14,7 +14,7 @@ from keras.backend import set_floatx
 from myNN import *
 from plotsUtil_batt import *
 
-set_floatx("float64")
+set_floatx("float32")
 
 from forwardPass import (
     from_param_list_to_str,
