+++ pinn_spm_param/util/_rescale.py
@@ -4,7 +4,7 @@ import tensorflow as tf
 from conditionalDecorator import conditional_decorator
 from keras.backend import set_floatx
 
-set_floatx("float64")
+set_floatx("float32")
 
 # Read command line arguments
 args = argument.initArg()
@@ -21,23 +21,23 @@ def rescalePhie(self, phie, t, deg_i0_a, deg_ds_c):
 
     if self.use_hnntime:
         phie_start = self.get_phie_hnntime(deg_i0_a_reshape, deg_ds_c_reshape)
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape - self.hnntime_val) / self.hard_IC_timescale
         )
     else:
         phie_start = self.get_phie0(deg_i0_a_reshape)
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape) / self.hard_IC_timescale
         )
 
-    offset = np.float64(0.0)
+    offset = np.float32(0.0)
     phie_nn = phie
     if self.use_hnn:
         phie_hnn = self.get_phie_hnn(
             t_reshape, deg_i0_a_reshape, deg_ds_c_reshape
         )
         offset = phie_hnn - phie_start
-        resc_phie *= np.float64(0.1)
+        resc_phie *= np.float32(0.1)
 
     return (resc_phie * phie_nn + offset) * timeDistance + phie_start
 
@@ -55,23 +55,23 @@ def rescalePhis_c(self, phis_c, t, deg_i0_a, deg_ds_c):
         phis_c_start = self.get_phis_c_hnntime(
             deg_i0_a_reshape, deg_ds_c_reshape
         )
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape - self.hnntime_val) / self.hard_IC_timescale
         )
     else:
         phis_c_start = self.get_phis_c0(deg_i0_a_reshape)
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape) / self.hard_IC_timescale
         )
 
-    offset = np.float64(0.0)
+    offset = np.float32(0.0)
     phis_c_nn = phis_c
     if self.use_hnn:
         phis_c_hnn = self.get_phis_c_hnn(
             t_reshape, deg_i0_a_reshape, deg_ds_c_reshape
         )
         offset = phis_c_hnn - phis_c_start
-        resc_phis_c *= np.float64(0.1)
+        resc_phis_c *= np.float32(0.1)
 
     return (resc_phis_c * phis_c_nn + offset) * timeDistance + phis_c_start
 
@@ -90,30 +90,30 @@ def rescaleCs_a(self, cs_a, t, r, deg_i0_a, deg_ds_c, clip=True):
         cs_a_start = self.get_cs_a_hnntime(
             r_reshape, deg_i0_a_reshape, deg_ds_c_reshape
         )
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape - self.hnntime_val) / self.hard_IC_timescale
         )
     else:
         cs_a_start = self.cs_a0
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape) / self.hard_IC_timescale
         )
     resc_cs_a = -cs_a_start
 
-    offset = np.float64(0.0)
+    offset = np.float32(0.0)
     if self.use_hnn:
         cs_a_hnn = self.get_cs_a_hnn(
             t_reshape, r_reshape, deg_i0_a_reshape, deg_ds_c_reshape
         )
         offset = cs_a_hnn - cs_a_start
-        cs_a_nn = cs_a * np.float64(0.01)
+        cs_a_nn = cs_a * np.float32(0.01)
     else:
         cs_a_nn = tf.math.sigmoid(cs_a)
 
     if clip:
         return tf.clip_by_value(
             (resc_cs_a * cs_a_nn + offset) * timeDistance + cs_a_start,
-            np.float64(0.0),
+            np.float32(0.0),
             self.params["csanmax"],
         )
     else:
@@ -134,30 +134,30 @@ def rescaleCs_c(self, cs_c, t, r, deg_i0_a, deg_ds_c, clip=True):
         cs_c_start = self.get_cs_c_hnntime(
             r_reshape, deg_i0_a_reshape, deg_ds_c_reshape
         )
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape - self.hnntime_val) / self.hard_IC_timescale
         )
     else:
         cs_c_start = self.cs_c0
-        timeDistance = np.float64(1.0) - tf.exp(
+        timeDistance = np.float32(1.0) - tf.exp(
             -(t_reshape) / self.hard_IC_timescale
         )
     resc_cs_c = self.params["cscamax"] - cs_c_start
 
-    offset = np.float64(0.0)
+    offset = np.float32(0.0)
     if self.use_hnn:
         cs_c_hnn = self.get_cs_c_hnn(
             t_reshape, r_reshape, deg_i0_a_reshape, deg_ds_c_reshape
         )
         offset = cs_c_hnn - cs_c_start
-        cs_c_nn = cs_c * np.float64(0.01)
+        cs_c_nn = cs_c * np.float32(0.01)
     else:
         cs_c_nn = tf.math.sigmoid(cs_c)
 
     if clip:
         return tf.clip_by_value(
             (resc_cs_c * cs_c_nn + offset) * timeDistance + cs_c_start,
-            np.float64(0.0),
+            np.float32(0.0),
             self.params["cscamax"],
         )
     else:
@@ -167,9 +167,9 @@ def rescaleCs_c(self, cs_c, t, r, deg_i0_a, deg_ds_c, clip=True):
 def get_phie0(self, deg_i0_a):
     i0_a = self.params["i0_a"](
         self.params["cs_a0"]
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         self.params["ce0"]
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         self.params["T"],
         self.params["alpha_a"],
         self.params["csanmax"],
@@ -189,9 +189,9 @@ def get_phie0(self, deg_i0_a):
 def get_phis_c0(self, deg_i0_a):
     i0_a = self.params["i0_a"](
         self.params["cs_a0"]
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         self.params["ce0"]
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         self.params["T"],
         self.params["alpha_a"],
         self.params["csanmax"],
@@ -219,7 +219,7 @@ def get_phie_hnn(self, t, deg_i0_a, deg_ds_c):
             self.hnn.model(
                 [
                     t / self.hnn.params["rescale_T"],
-                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float64),
+                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float32),
                     self.hnn.rescale_param(
                         self.fix_param(
                             deg_i0_a, self.hnn_params[self.hnn.ind_deg_i0_a]
@@ -244,7 +244,7 @@ def get_phie_hnn(self, t, deg_i0_a, deg_ds_c):
             self.hnn.model(
                 [
                     t / self.hnn.params["rescale_T"],
-                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float64),
+                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float32),
                     self.hnn.rescale_param(deg_i0_a, self.hnn.ind_deg_i0_a),
                     self.hnn.rescale_param(deg_ds_c, self.hnn.ind_deg_ds_c),
                 ],
@@ -261,9 +261,9 @@ def get_phie_hnntime(self, deg_i0_a, deg_ds_c):
         self.hnntime.model(
             [
                 self.hnntime_val
-                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64)
+                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32)
                 / self.hnntime.params["rescale_T"],
-                tf.zeros(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+                tf.zeros(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
                 self.hnntime.rescale_param(
                     deg_i0_a, self.hnntime.ind_deg_i0_a
                 ),
@@ -274,7 +274,7 @@ def get_phie_hnntime(self, deg_i0_a, deg_ds_c):
             training=False,
         )[self.hnntime.ind_phie],
         self.hnntime_val
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         deg_i0_a,
         deg_ds_c,
     )
@@ -287,7 +287,7 @@ def get_phis_c_hnn(self, t, deg_i0_a, deg_ds_c):
             self.hnn.model(
                 [
                     t / self.hnn.params["rescale_T"],
-                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float64),
+                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float32),
                     self.hnn.rescale_param(
                         self.fix_param(
                             deg_i0_a, self.hnn_params[self.hnn.ind_deg_i0_a]
@@ -312,7 +312,7 @@ def get_phis_c_hnn(self, t, deg_i0_a, deg_ds_c):
             self.hnn.model(
                 [
                     t / self.hnn.params["rescale_T"],
-                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float64),
+                    tf.zeros(tf.shape(t), dtype=tf.dtypes.float32),
                     self.hnn.rescale_param(deg_i0_a, self.hnn.ind_deg_i0_a),
                     self.hnn.rescale_param(deg_ds_c, self.hnn.ind_deg_ds_c),
                 ],
@@ -329,9 +329,9 @@ def get_phis_c_hnntime(self, deg_i0_a, deg_ds_c):
         self.hnntime.model(
             [
                 self.hnntime_val
-                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64)
+                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32)
                 / self.hnntime.params["rescale_T"],
-                tf.zeros(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+                tf.zeros(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
                 self.hnntime.rescale_param(
                     deg_i0_a, self.hnntime.ind_deg_i0_a
                 ),
@@ -342,7 +342,7 @@ def get_phis_c_hnntime(self, deg_i0_a, deg_ds_c):
             training=False,
         )[self.hnntime.ind_phis_c],
         self.hnntime_val
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         deg_i0_a,
         deg_ds_c,
     )
@@ -399,7 +399,7 @@ def get_cs_a_hnntime(self, r, deg_i0_a, deg_ds_c):
         self.hnntime.model(
             [
                 self.hnntime_val
-                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64)
+                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32)
                 / self.hnntime.params["rescale_T"],
                 r / self.hnn.params["rescale_R"],
                 self.hnntime.rescale_param(
@@ -412,7 +412,7 @@ def get_cs_a_hnntime(self, r, deg_i0_a, deg_ds_c):
             training=False,
         )[self.hnntime.ind_cs_a],
         self.hnntime_val
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         r / self.hnn.params["rescale_R"],
         deg_i0_a,
         deg_ds_c,
@@ -470,7 +470,7 @@ def get_cs_c_hnntime(self, r, deg_i0_a, deg_ds_c):
         self.hnntime.model(
             [
                 self.hnntime_val
-                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64)
+                * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32)
                 / self.hnntime.params["rescale_T"],
                 r / self.hnn.params["rescale_R"],
                 self.hnntime.rescale_param(
@@ -483,7 +483,7 @@ def get_cs_c_hnntime(self, r, deg_i0_a, deg_ds_c):
             training=False,
         )[self.hnntime.ind_cs_c],
         self.hnntime_val
-        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float64),
+        * tf.ones(tf.shape(deg_i0_a), dtype=tf.dtypes.float32),
         r / self.hnn.params["rescale_R"],
         deg_i0_a,
         deg_ds_c,
@@ -495,7 +495,7 @@ def rescale_param(self, param, ind):
 
 
 def fix_param(self, param, param_val):
-    return param_val * tf.ones(tf.shape(param), dtype=tf.dtypes.float64)
+    return param_val * tf.ones(tf.shape(param), dtype=tf.dtypes.float32)
 
 
 def unrescale_param(self, param_rescaled, ind):
